feat: extract HMAC signature verification to a dedicated module and update webhook route to use POST method
